LayoutTests/imported/w3c/web-platform-tests/uievents/textInput/delete-expected.txt

Press Delete (Fn+Backspace for macOS) in each text field below.


abc

FAIL textInput: delete, <input> assert_equals: expected "ac" but got "abc"
FAIL textInput: delete, <textarea> assert_equals: expected "ac" but got "abc"
FAIL textInput: delete, <div contenteditable="true"> assert_equals: expected "ac" but got "abc"
